requirements.txt
# Core
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bootstrap5
Flask-Mail==0.9.1
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.19
Flask-Migrate==4.0.4

# Security
Werkzeug==2.3.6
email-validator==2.0.0.post2

# Development
Flask-DebugToolbar==0.13.1